At this point seeing life in the suite with use of geosmeta and up to calling launchruns. Replicating some code in tasks because not yet got elegant dependencies, Note both graphs begin togetehr - so the first run of the periodic one can be  before the studydirectory is made
Note the -C is needed because HOME and ~ get redefined
Attempt to use A2R as a list failed with unrecognised arguments - may be to do with use of quotes - so doing invidiual calls to send request

Currently seetign the study name in two places in the rose-suite.conf. Must be able to do better.... can -D be used to define study name?
The state file should be set using the python utility not just an echo, but I was in a rush to test.